+++ prisma/migrations/20260902175303_add_likes/migration.sql
@@ -1,0 +1,31 @@
+-- CreateTable
+CREATE TABLE "Like" (
+    "id" TEXT NOT NULL,
+    "userId" TEXT NOT NULL,
+    "logEntryId" TEXT,
+    "listId" TEXT,
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+
+    CONSTRAINT "Like_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE INDEX "Like_logEntryId_idx" ON "Like"("logEntryId");
+
+-- CreateIndex
+CREATE INDEX "Like_listId_idx" ON "Like"("listId");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Like_userId_logEntryId_key" ON "Like"("userId", "logEntryId");
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Like_userId_listId_key" ON "Like"("userId", "listId");
+
+-- AddForeignKey
+ALTER TABLE "Like" ADD CONSTRAINT "Like_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User"("id") ON DELETE CASCADE ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "Like" ADD CONSTRAINT "Like_logEntryId_fkey" FOREIGN KEY ("logEntryId") REFERENCES "LogEntry"("id") ON DELETE CASCADE ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "Like" ADD CONSTRAINT "Like_listId_fkey" FOREIGN KEY ("listId") REFERENCES "List"("id") ON DELETE CASCADE ON UPDATE CASCADE;